python/telega/requirements.txt
aiogram==3.21.0
selenium==4.34.2
undetected-chromedriver==3.5.5
webdriver-manager==4.0.2
pillow==11.3.0
requests==2.32.4
python-dotenv==1.1.1
aiohttp==3.12.14